Implementation moved to a separate file
